+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Default route to `roles/list
+
+### Removed
+- Tabs
+
 ## [0.1.2] - 2021-09-07
 
 ### Fixed